Separate ServiceManager tests into trait
Pushing them to a trait allows executing them on the
AbstractPluginManager as well, ensuring feature parity.

This patch also incorporates other feedback as provided by @Maks3w.
