+++ test/TestAsset/LenientPluginManager.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace ZendTest\ServiceManager\TestAsset;
+
+use Zend\ServiceManager\AbstractPluginManager;
+
+class LenientPluginManager extends AbstractPluginManager
+{
+    /**
+     * Allow anything to be considered valid.
+     */
+    public function validate($instance)
+    {
+        return;
+    }
+}
